+++ axum/src/response/sse.rs
@@ -163,9 +163,9 @@ pub struct Event {
 }
 
 impl Event {
-    /// Set the event's data data field(s) (`data:<content>`)
+    /// Set the event's data data field(s) (`data: <content>`)
     ///
-    /// Newlines in `data` will automatically be broken across `data:` fields.
+    /// Newlines in `data` will automatically be broken across `data: ` fields.
     ///
     /// This corresponds to [`MessageEvent`'s data field].
     ///
@@ -194,7 +194,7 @@ impl Event {
         self
     }
 
-    /// Set the event's data field to a value serialized as unformatted JSON (`data:<content>`).
+    /// Set the event's data field to a value serialized as unformatted JSON (`data: <content>`).
     ///
     /// This corresponds to [`MessageEvent`'s data field].
     ///
@@ -212,7 +212,7 @@ impl Event {
             panic!("Called `EventBuilder::json_data` multiple times");
         }
 
-        self.buffer.extend_from_slice(b"data:");
+        self.buffer.extend_from_slice(b"data: ");
         serde_json::to_writer((&mut self.buffer).writer(), &data).map_err(axum_core::Error::new)?;
         self.buffer.put_u8(b'\n');
 
@@ -350,10 +350,7 @@ impl Event {
         );
         self.buffer.extend_from_slice(name.as_bytes());
         self.buffer.put_u8(b':');
-        // Prevent values that start with spaces having that space stripped
-        if value.starts_with(b" ") {
-            self.buffer.put_u8(b' ');
-        }
+        self.buffer.put_u8(b' ');
         self.buffer.extend_from_slice(value);
         self.buffer.put_u8(b'\n');
     }
@@ -374,25 +371,19 @@ impl EventFlags {
     const HAS_ID: Self = Self::from_bits(0b1000);
 
     const fn bits(&self) -> u8 {
-        let bits = self;
-        bits.0
+        self.0
     }
 
     const fn from_bits(bits: u8) -> Self {
-        let bits = bits;
         Self(bits)
     }
 
     const fn contains(&self, other: Self) -> bool {
-        let same = self;
-        let other = other;
-        same.bits() & other.bits() == other.bits()
+        self.bits() & other.bits() == other.bits()
     }
 
     fn insert(&mut self, other: Self) {
-        let same = self;
-        let other = other;
-        *same = Self::from_bits(same.bits() | other.bits());
+        *self = Self::from_bits(self.bits() | other.bits());
     }
 }
 
@@ -530,7 +521,7 @@ mod tests {
     #[test]
     fn leading_space_is_not_stripped() {
         let no_leading_space = Event::default().data("\tfoobar");
-        assert_eq!(&*no_leading_space.finalize(), b"data:\tfoobar\n\n");
+        assert_eq!(&*no_leading_space.finalize(), b"data: \tfoobar\n\n");
 
         let leading_space = Event::default().data(" foobar");
         assert_eq!(&*leading_space.finalize(), b"data:  foobar\n\n");
@@ -679,7 +670,7 @@ mod tests {
     }
 
     #[test]
-    fn memchr_spliting() {
+    fn memchr_splitting() {
         assert_eq!(
             memchr_split(2, &[]).collect::<Vec<_>>(),
             [&[]] as [&[u8]; 1]