+++ axum/src/routing/path_router.rs
@@ -1,4 +1,4 @@
-use crate::extract::Request;
+use crate::extract::{nested_path::SetNestedPath, Request};
 use axum_core::response::IntoResponse;
 use matchit::MatchError;
 use std::{borrow::Cow, collections::HashMap, convert::Infallible, fmt, sync::Arc};
@@ -53,8 +53,6 @@ where
 
         validate_path(path)?;
 
-        let id = self.next_route_id();
-
         let endpoint = if let Some((route_id, Endpoint::MethodRouter(prev_method_router))) = self
             .node
             .path_to_route_id
@@ -74,6 +72,7 @@ where
             Endpoint::MethodRouter(method_router)
         };
 
+        let id = self.next_route_id();
         self.set_node(path, id)?;
         self.routes.insert(id, endpoint);
 
@@ -163,10 +162,10 @@ where
 
     pub(super) fn nest(
         &mut self,
-        path: &str,
+        path_to_nest_at: &str,
         router: PathRouter<S, IS_FALLBACK>,
     ) -> Result<(), Cow<'static, str>> {
-        let prefix = validate_nest_path(path);
+        let prefix = validate_nest_path(path_to_nest_at);
 
         let PathRouter {
             routes,
@@ -182,7 +181,11 @@ where
 
             let path = path_for_nested_route(prefix, inner_path);
 
-            match endpoint.layer(StripPrefix::layer(prefix)) {
+            let layer = (
+                StripPrefix::layer(prefix),
+                SetNestedPath::layer(path_to_nest_at),
+            );
+            match endpoint.layer(layer) {
                 Endpoint::MethodRouter(method_router) => {
                     self.route(&path, method_router)?;
                 }
@@ -195,13 +198,17 @@ where
         Ok(())
     }
 
-    pub(super) fn nest_service<T>(&mut self, path: &str, svc: T) -> Result<(), Cow<'static, str>>
+    pub(super) fn nest_service<T>(
+        &mut self,
+        path_to_nest_at: &str,
+        svc: T,
+    ) -> Result<(), Cow<'static, str>>
     where
         T: Service<Request, Error = Infallible> + Clone + Send + 'static,
         T::Response: IntoResponse,
         T::Future: Send + 'static,
     {
-        let path = validate_nest_path(path);
+        let path = validate_nest_path(path_to_nest_at);
         let prefix = path;
 
         let path = if path.ends_with('/') {
@@ -210,7 +217,11 @@ where
             format!("{path}/*{NEST_TAIL_PARAM}")
         };
 
-        let endpoint = Endpoint::Route(Route::new(StripPrefix::new(svc, prefix)));
+        let layer = (
+            StripPrefix::layer(prefix),
+            SetNestedPath::layer(path_to_nest_at),
+        );
+        let endpoint = Endpoint::Route(Route::new(layer.layer(svc)));
 
         self.route_endpoint(&path, endpoint.clone())?;