+++ axum/src/routing/tests/mod.rs
@@ -20,10 +20,11 @@ use std::{
     task::{Context, Poll},
     time::Duration,
 };
-use tower::{
-    service_fn, timeout::TimeoutLayer, util::MapResponseLayer, ServiceBuilder, ServiceExt,
+use tower::{service_fn, util::MapResponseLayer, ServiceExt};
+use tower_http::{
+    limit::RequestBodyLimitLayer, timeout::TimeoutLayer,
+    validate_request::ValidateRequestHeaderLayer,
 };
-use tower_http::{limit::RequestBodyLimitLayer, validate_request::ValidateRequestHeaderLayer};
 use tower_service::Service;
 
 mod fallback;
@@ -169,7 +170,6 @@ async fn routing_between_services() {
 
 #[crate::test]
 async fn middleware_on_single_route() {
-    use tower::ServiceBuilder;
     use tower_http::{compression::CompressionLayer, trace::TraceLayer};
 
     async fn handle(_: Request) -> &'static str {
@@ -178,12 +178,7 @@ async fn middleware_on_single_route() {
 
     let app = Router::new().route(
         "/",
-        get(handle.layer(
-            ServiceBuilder::new()
-                .layer(TraceLayer::new_for_http())
-                .layer(CompressionLayer::new())
-                .into_inner(),
-        )),
+        get(handle.layer((TraceLayer::new_for_http(), CompressionLayer::new()))),
     );
 
     let client = TestClient::new(app);
@@ -299,13 +294,7 @@ async fn wildcard_sees_whole_url() {
 async fn middleware_applies_to_routes_above() {
     let app = Router::new()
         .route("/one", get(std::future::pending::<()>))
-        .layer(
-            ServiceBuilder::new()
-                .layer(HandleErrorLayer::new(|_: BoxError| async move {
-                    StatusCode::REQUEST_TIMEOUT
-                }))
-                .layer(TimeoutLayer::new(Duration::new(0, 0))),
-        )
+        .layer(TimeoutLayer::new(Duration::ZERO))
         .route("/two", get(|| async {}));
 
     let client = TestClient::new(app);
@@ -730,6 +719,75 @@ async fn changing_the_default_limit() {
 }
 
 #[crate::test]
+async fn changing_the_default_limit_differently_on_different_routes() {
+    let limit1 = 2;
+    let limit2 = 10;
+
+    let app = Router::new()
+        .route(
+            "/limit1",
+            post(|_: Bytes| async {}).layer(DefaultBodyLimit::max(limit1)),
+        )
+        .route(
+            "/limit2",
+            post(|_: Bytes| async {}).layer(DefaultBodyLimit::max(limit2)),
+        )
+        .route("/default", post(|_: Bytes| async {}));
+
+    let client = TestClient::new(app);
+
+    let res = client
+        .post("/limit1")
+        .body(reqwest::Body::from("a".repeat(limit1)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::OK);
+
+    let res = client
+        .post("/limit1")
+        .body(reqwest::Body::from("a".repeat(limit2)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
+
+    let res = client
+        .post("/limit2")
+        .body(reqwest::Body::from("a".repeat(limit1)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::OK);
+
+    let res = client
+        .post("/limit2")
+        .body(reqwest::Body::from("a".repeat(limit2)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::OK);
+
+    let res = client
+        .post("/limit2")
+        .body(reqwest::Body::from("a".repeat(limit1 + limit2)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
+
+    let res = client
+        .post("/default")
+        .body(reqwest::Body::from("a".repeat(limit1 + limit2)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::OK);
+
+    let res = client
+        .post("/default")
+        // `DEFAULT_LIMIT` is 2mb so make a body larger than that
+        .body(reqwest::Body::from("a".repeat(3_000_000)))
+        .send()
+        .await;
+    assert_eq!(res.status(), StatusCode::PAYLOAD_TOO_LARGE);
+}
+
+#[crate::test]
 async fn limited_body_with_streaming_body() {
     const LIMIT: usize = 3;
 
@@ -878,7 +936,7 @@ async fn state_isnt_cloned_too_much() {
                         .filter(|line| line.contains("axum") || line.contains("./src"))
                         .collect::<Vec<_>>()
                         .join("\n");
-                    println!("AppState::Clone:\n===============\n{}\n", bt);
+                    println!("AppState::Clone:\n===============\n{bt}\n");
                     COUNT.fetch_add(1, Ordering::SeqCst);
                 }
             }