+++ axum/src/serve.rs
@@ -4,7 +4,10 @@ use std::{convert::Infallible, io, net::SocketAddr};
 
 use axum_core::{body::Body, extract::Request, response::Response};
 use futures_util::{future::poll_fn, FutureExt};
-use hyper::server::conn::http1;
+use hyper_util::{
+    rt::{TokioExecutor, TokioIo},
+    server::conn::auto::Builder,
+};
 use tokio::net::{TcpListener, TcpStream};
 use tower_service::Service;
 
@@ -13,7 +16,7 @@ use tower_service::Service;
 /// This method of running a service is intentionally simple and doesn't support any configuration.
 /// Use hyper or hyper-util if you need configuration.
 ///
-/// It only supports HTTP/1.
+/// It supports both HTTP/1 as well as HTTP/2.
 ///
 /// # Examples
 ///
@@ -122,15 +125,14 @@ where
         });
 
         tokio::task::spawn(async move {
-            match http1::Builder::new()
-                .serve_connection(tcp_stream, service)
-                // for websockets
-                .with_upgrades()
+            match Builder::new(TokioExecutor::new())
+                // upgrades needed for websockets
+                .serve_connection_with_upgrades(tcp_stream.into_inner(), service)
                 .await
             {
                 Ok(()) => {}
                 Err(_err) => {
-                    // This error only appears when  the client doesn't send a request and
+                    // This error only appears when the client doesn't send a request and
                     // terminate the connection.
                     //
                     // If client sends one request then terminate connection whenever, it doesn't